backend/requirements.txt
devtools
fastapi
google-generativeai
opencv-python
Pillow
psycopg2-binary
pydantic
python-dotenv
sqlalchemy
uvicorn
